requirements.txt
fastf1>=3.1.0
rich>=13.0.0
Pillow>=10.0.0 